refactor(commands/handlers/behaviors): 重命名部分文件以符合 kekab-case
