+++ src/features/commands/handlers/behaviors/break-block.ts
