+++ src/features/commands/handlers/index.ts
@@ -1,12 +1,12 @@
 export * from './meta/help';
-export * from './meta/list';
+export * from './meta/list-simulated-players';
 export * from './meta/location';
 export * from './meta/rename';
 export * from './meta/setting';
-export * from './meta/showCommandsList';
+export * from './meta/list-commands';
 export * from './meta/time';
 
-export * from './lifecycle/chatSpawn';
+export * from './lifecycle/spawn';
 export * from './lifecycle/disconnect';
 export * from './lifecycle/respawn';
 
@@ -17,6 +17,6 @@ export * from './inventory/offhand-swap';
 export * from './inventory/recycle';
 
 export * from './behaviors/behaviors';
-export * from './behaviors/breakBlock';
+export * from './behaviors/break-block';
 
 export * from './tps/indicator'
