+++ src/features/commands/handlers/lifecycle/spawn.ts
